pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ztare"
version = "0.1.0"
description = "ZTARE: zero-trust adversarial validation for claims, theses, and strategic logic"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-genai>=1.0.0",
  "anthropic>=0.25.0",
  "openai>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
